backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
python-docx==1.1.0
tqdm==4.66.1
sentence-transformers==2.3.1
faiss-cpu>=1.8.0
python-multipart==0.0.6

